make cat movement smoother
I had an idea this morning that we were forcing the cat to move the max
speed all of the time if the cat moved, which was making it a bit jumpy.

I cleaned it up by making the cat move *either* the total distance to
the ball *or* the max speed, whichever is less (using Clamp).  This way
if the ball less than maxSpeed away, he still moves towards it, but in
a smaller jump.

Because I'm having the cat move the distance to the ball (rather than
maxSpeed), the distance contains the direction (+/-) and we don't need
direction anymore.